Extract two internal query classes from EduProgramProfileQueryBase, mark it abstract base class
