+++ R7.University.EduProgramProfileDirectory/R7.University.EduProgramProfileDirectory.csproj
@@ -92,6 +92,7 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ViewModels\EduProgramProfileDirectoryDocumentsViewModel.cs" />
     <Compile Include="ViewModels\EduProgramProfileDirectoryEduFormsViewModel.cs" />
+    <Compile Include="Queries\EduProgramProfileQuery.cs" />
   </ItemGroup>
   <ItemGroup>
     <Folder Include="Components\" />
