+++ R7.University.EduProgramProfileDirectory/ViewEduProgramProfileDirectory.ascx.cs
@@ -42,6 +42,7 @@ using R7.University.ModelExtensions;
 using R7.University.Models;
 using R7.University.Queries;
 using R7.University.ViewModels;
+using R7.University.EduProgramProfileDirectory.Queries;
 
 namespace R7.University.EduProgramProfileDirectory
 {
@@ -112,7 +113,7 @@ namespace R7.University.EduProgramProfileDirectory
             var viewModel = new EduProgramProfileDirectoryEduFormsViewModel ();
             var indexer = new ViewModelIndexer (1);
 
-            var eduProgramProfiles = new EduProgramProfileQueryBase (ModelContext).ListByEduLevelsWithEduForms (Settings.EduLevels);
+            var eduProgramProfiles = new EduProgramProfileQuery (ModelContext).ListByEduLevelsWithEduForms (Settings.EduLevels);
                
             viewModel.EduProgramProfiles = new IndexedEnumerable<EduProgramProfileObrnadzorEduFormsViewModel> (indexer,
                 eduProgramProfiles.Select (epp => new EduProgramProfileObrnadzorEduFormsViewModel (epp, viewModel, indexer))
@@ -126,7 +127,7 @@ namespace R7.University.EduProgramProfileDirectory
             var viewModel = new EduProgramProfileDirectoryDocumentsViewModel ();
             var indexer = new ViewModelIndexer (1);
 
-            var eduProgramProfiles = new EduProgramProfileQueryBase (ModelContext).ListByEduLevelsWithDocuments (Settings.EduLevels);
+            var eduProgramProfiles = new EduProgramProfileQuery (ModelContext).ListByEduLevelsWithDocuments (Settings.EduLevels);
 
             viewModel.EduProgramProfiles = new IndexedEnumerable<EduProgramProfileObrnadzorDocumentsViewModel> (indexer,
                 eduProgramProfiles.Select (epp => new EduProgramProfileObrnadzorDocumentsViewModel (epp, viewModel, indexer))
